+++ Example/Example/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
