+++ Example/Example/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
