+++ Example/Example/Resources/Assets.xcassets/Contents.json
