+++ Sources/AppEntry/app.scene.swift
@@ -15,7 +15,7 @@ struct MainScene: Scene {
 
 
 
-extension ApplicationContext.State {
+public extension ApplicationContext.State {
     @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
     var mainScene: WindowGroup<AnyView>? {
         get {
