+++ Sources/AppService/service.module.swift
@@ -1,8 +1,7 @@
 import Foundation
 
 public protocol ServiceModule {
-    @MainActor
-    init(_ context: ApplicationContext)
+   
 }
 
 public extension ApplicationContext {
